requirements/default.txt
cachetools==4.2.2
gspread_pandas==3.2.2
httplib2==0.19.1
oauth2client==4.1.3
gspread==5.4.0
numpy== 1.22.0
pandas==1.3.3
brotli==1.0.9
Cython==0.29.24
docutils>=0.15.2
google_api_python_client==2.18.0
importlib_metadata==4.8.1
keyring==23.1.0
lockfile==0.12.2
lxml==4.9.1
mock==4.0.3
ordereddict==1.1
protobuf==3.20.2
pyOpenSSL==20.0.1
simplejson==3.17.5
toml==0.10.2
unicodedataplus==14.0.0.post2
usercustomize==1.0.0
zipp==3.6.0